packages/backend-python/pyproject.toml
[tool.poetry]
name = "helios-backend"
version = "1.0.0"
description = "Helios backend - Agent runtime and orchestration in Python"
authors = ["Helios Team"]
readme = "README.md"
packages = [{include = "helios"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
python-socketio = "^5.11.0"
asyncpg = "^0.29.0"
redis = "^5.0.1"
python-dotenv = "^1.0.0"
pydantic = "^2.6.0"
pydantic-settings = "^2.2.1"
loguru = "^0.7.2"
httpx = "^0.26.0"
langchain = "^0.1.0"
langgraph = "^0.0.20"
openai = "^1.12.0"
anthropic = "^0.18.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"
black = "^24.1.0"
ruff = "^0.2.0"
mypy = "^1.8.0"
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]

[tool.black]
line-length = 120
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
